Reorganize modules for cff parsing
A lot of code was moved out of the tables::cff module to make the file more
readable. Instead a new top-level cff module was introduced with some
submodules that export helper structs and functions that are needed for
cff parsing.

This reorganization will also be needed for truetype table data at some
point.